+++ src/motion_lib/include/motion_lib/differential_model/base_differential_model.hpp
@@ -23,5 +23,5 @@ public:
    * Nm
    */
   virtual common_lib::structures::Wheels calculateTorqueDistribution(
-      float input_torque, const common_lib::structures::Wheels& wheel_speeds) const = 0;
+      double input_torque, const common_lib::structures::Wheels& wheel_speeds) const = 0;
 };
